Remove MetaUtils
