+++ molgenis-data/src/main/java/org/molgenis/data/util/EntityTypeUtils.java
@@ -1,6 +1,15 @@
 package org.molgenis.data.util;
 
 import static java.util.stream.StreamSupport.stream;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.ATTRIBUTES;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.BACKEND;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.DESCRIPTION;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.EXTENDS;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.ID;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.IS_ABSTRACT;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.LABEL;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.PACKAGE;
+import static org.molgenis.data.meta.model.EntityTypeMetadata.TAGS;
 import static org.molgenis.data.meta.model.Package.PACKAGE_SEPARATOR;
 import static org.molgenis.data.util.PackageUtils.isSystemPackage;
 
@@ -326,5 +335,19 @@ public class EntityTypeUtils {
 
   public static boolean isSystemEntity(EntityType entityType) {
     return isSystemPackage(entityType.getPackage());
+  }
+
+  public static Fetch getEntityTypeFetch() {
+    // TODO simplify fetch creation (in this case *all* attributes and expand xref/mrefs)
+    return new Fetch()
+        .field(ID)
+        .field(PACKAGE)
+        .field(LABEL)
+        .field(DESCRIPTION)
+        .field(ATTRIBUTES)
+        .field(IS_ABSTRACT)
+        .field(EXTENDS)
+        .field(TAGS)
+        .field(BACKEND);
   }
 }
