+++ molgenis-data/src/test/java/org/molgenis/data/meta/MetaDataServiceImplTest.java
@@ -24,7 +24,7 @@ import static org.molgenis.data.meta.model.EntityTypeMetadata.IS_ABSTRACT;
 import static org.molgenis.data.meta.model.PackageMetadata.PACKAGE;
 import static org.molgenis.data.meta.model.PackageMetadata.PARENT;
 import static org.molgenis.data.meta.model.TagMetadata.TAG;
-import static org.molgenis.data.util.MetaUtils.getEntityTypeFetch;
+import static org.molgenis.data.util.EntityTypeUtils.getEntityTypeFetch;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
